+++ src/rgw/CMakeLists.txt
@@ -42,6 +42,10 @@ set(librgw_common_srcs
   services/svc_finisher.cc
   services/svc_bucket.cc
   services/svc_cls.cc
+  services/svc_mdlog.cc
+  services/svc_meta.cc
+  services/svc_meta_be.cc
+  services/svc_meta_be_sobj.cc
   services/svc_notify.cc
   services/svc_quota.cc
   services/svc_sync_modules.cc
